fix(rapport-responsables-donnees): simplifier le contenu rapport et aligner sur le template 84 CH-197
- Nouveau type ContenuRapportResponsableDonnees avec seulement nom_chantier, id_chantier, indicateursNonMisAJour, nombreIndicateursNonMisAJour
- genererParametresRapportResponsableDonnees déplacée dans ParametresEnvoieEmailRapportResponsableDonnees.ts avec ses tests
- Template Brevo 4 → 84, payload simplifié (suppression conseillerEmail et texteIntro)
- Mise à jour du schéma Zod du repository et de tous les tests associés
