+++ apps/pilote-ppg/src/server/chantiers/__tests__/usecases/EnvoyerLesRapportsResponsablesDonneesUseCase.unit.test.ts
@@ -17,22 +17,12 @@ function creerRapportTest(params: {
         {
           nom_chantier: "Chantier 197",
           id_chantier: "CH-197",
-          nombre_propositions: "",
-          conseiller_email: "conseiller@test.com",
-          afficherSectionPropositions: false,
-          indicateursPropositions: [],
-          afficherSectionMajIndicateur: true,
           indicateursNonMisAJour: [
             { id: "IND-001", nom: "Indicateur 1", mailles: ["NAT"] },
           ],
           nombreIndicateursNonMisAJour: "1 indicateur à mettre à jour",
-          afficherSectionParametrage: false,
-          indicateursAParametrer: [],
-          nombreIndicateursAParametrer: "",
         },
       ],
-      conseillerEmail: "conseiller@test.com",
-      texteIntro: "votre chantier prioritaire",
     },
     statutEnvoi: "CREE",
     dateCreation: new Date("2026-02-03"),
